almost json parser and config
Adding this example to the fluentbit examples repo. This situation comes in AWS environments where the logs are almost json, but have some text in the beginning of the log record making it an incomplete json payload.

This parser will strip out the beginning text and write the json payload to the message field so it can automatically be parsed by NR.